# backend/src/openapi.json
{"openapi":"3.1.0","info":{"title":"Textbook RAG API","version":"0.1.0"},"paths":{"/api/v1/textbook/chapters":{"get":{"tags":["textbook"],"summary":"Get Textbook Chapters","description":"Get all textbook chapters","operationId":"get_textbook_chapters_api_v1_textbook_chapters_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TextbookChapterWithId"},"type":"array","title":"Response Get Textbook Chapters Api V1 Textbook Chapters Get"}}}}}}},"/api/v1/textbook/chapters/{chapter_id}":{"get":{"tags":["textbook"],"summary":"Get Textbook Chapter","description":"Get a specific textbook chapter","operationId":"get_textbook_chapter_api_v1_textbook_chapters__chapter_id__get","parameters":[{"name":"chapter_id","in":"path","required":true,"schema":{"type":"string","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextbookChapter"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/start":{"post":{"tags":["chat"],"summary":"Start Chat Session","description":"Start a new chat session (requires authentication)","operationId":"start_chat_session_api_v1_chat_start_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatSession"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/chat/{session_id}/query":{"post":{"tags":["chat"],"summary":"Chat Query","description":"Send a query to the RAG chatbot (requires authentication)","operationId":"chat_query_api_v1_chat__session_id__query_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatQueryWithSelection"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/rag":{"post":{"tags":["chat"],"summary":"Process Rag Chat","description":"Process user question with RAG assistant.\nProcess authenticated user questions and return answers based only on provided context.","operationId":"process_rag_chat_api_v1_chat_rag_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RAGChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RAGChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/rag/test":{"post":{"tags":["chat"],"summary":"Test Rag Chat With Auth Check","description":"Test endpoint that shows authentication status before processing.","operationId":"test_rag_chat_with_auth_check_api_v1_chat_rag_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RAGChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/textbook/search":{"post":{"tags":["search"],"summary":"Textbook Search","description":"Search textbook content","operationId":"textbook_search_api_v1_textbook_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cohere/config":{"get":{"tags":["cohere"],"summary":"Get Cohere Config","description":"Get Cohere integration configuration (requires authentication)","operationId":"get_cohere_config_api_v1_cohere_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CohereConfig"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/verify":{"post":{"tags":["auth"],"summary":"Verify Auth","description":"Verify if the provided token is valid and user is authenticated.","operationId":"verify_auth_api_v1_auth_verify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCheckRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCheckResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Authenticate user and return JWT token.\nThis is a placeholder implementation - in a real system, you'd validate against a database.","operationId":"login_api_v1_auth_login_post","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","description":"Authenticate user and return access token","operationId":"login_for_access_token_token_post","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthCheckRequest":{"properties":{"token":{"type":"string","title":"Token"}},"type":"object","required":["token"],"title":"AuthCheckRequest"},"AuthCheckResponse":{"properties":{"authenticated":{"type":"boolean","title":"Authenticated"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["authenticated","message"],"title":"AuthCheckResponse"},"AuthenticationStatus":{"properties":{"status":{"$ref":"#/components/schemas/AuthenticationStatusEnum"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Token"},"expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires At"}},"type":"object","required":["status"],"title":"AuthenticationStatus","description":"Represents whether the user is authenticated (authenticated | unauthenticated)"},"AuthenticationStatusEnum":{"type":"string","enum":["authenticated","unauthenticated"],"title":"AuthenticationStatusEnum"},"ChatMessage":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["role","content","timestamp"],"title":"ChatMessage"},"ChatQueryWithSelection":{"properties":{"message":{"type":"string","title":"Message"},"context":{"anyOf":[{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array"},{"type":"null"}],"title":"Context"},"selected_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Selected Text"}},"type":"object","required":["message"],"title":"ChatQueryWithSelection"},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response"},"sources":{"items":{"$ref":"#/components/schemas/Source"},"type":"array","title":"Sources"},"confidence_score":{"type":"number","title":"Confidence Score"},"is_hallucinated":{"type":"boolean","title":"Is Hallucinated"},"session_id":{"type":"string","title":"Session Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["response","sources","confidence_score","is_hallucinated","session_id","timestamp"],"title":"ChatResponse"},"ChatSession":{"properties":{"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"context_window":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Context Window","default":[]}},"type":"object","required":["id","created_at","updated_at"],"title":"ChatSession"},"ChatStatusEnum":{"type":"string","enum":["success","no-content","unauthorized"],"title":"ChatStatusEnum"},"CohereConfig":{"properties":{"model_name":{"type":"string","title":"Model Name","default":"command-r-plus"},"temperature":{"type":"number","title":"Temperature","default":0.3},"max_tokens":{"type":"integer","title":"Max Tokens","default":1000},"supports_streaming":{"type":"boolean","title":"Supports Streaming","default":true}},"type":"object","title":"CohereConfig","description":"Configuration for Cohere API"},"ContentBlock":{"properties":{"chapter_id":{"type":"string","title":"Chapter Id"},"content":{"type":"string","title":"Content"},"block_type":{"type":"string","title":"Block Type"},"position":{"type":"integer","title":"Position"},"embedding_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Embedding Id"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["chapter_id","content","block_type","position","id","created_at","updated_at"],"title":"ContentBlock"},"ContextContent":{"properties":{"type":{"$ref":"#/components/schemas/ContextContentTypeEnum"},"content":{"type":"string","minLength":1,"title":"Content","description":"The actual text content"},"source_document_id":{"type":"string","minLength":1,"title":"Source Document Id","description":"Identifier for the source document"},"section_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Reference"},"retrieval_metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Retrieval Metadata"}},"type":"object","required":["type","content","source_document_id"],"title":"ContextContent","description":"Authorized text that serves as the source of truth for responses (user-selected text OR retrieved book chunks)"},"ContextContentTypeEnum":{"type":"string","enum":["user-selected","retrieved-excerpt"],"title":"ContextContentTypeEnum"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RAGChatRequest":{"properties":{"question":{"type":"string","title":"Question"},"context":{"$ref":"#/components/schemas/ContextContent"},"auth_status":{"anyOf":[{"$ref":"#/components/schemas/AuthenticationStatus"},{"type":"null"}]},"mode":{"type":"string","pattern":"^(selection-restricted|rag-mode)$","title":"Mode"}},"type":"object","required":["question","context","mode"],"title":"RAGChatRequest"},"RAGChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"status":{"$ref":"#/components/schemas/ChatStatusEnum"},"context_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context Used"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["answer","status","timestamp"],"title":"RAGChatResponse"},"SearchRequest":{"properties":{"query":{"type":"string","title":"Query"},"top_k":{"type":"integer","title":"Top K","default":5}},"type":"object","required":["query"],"title":"SearchRequest"},"SearchResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/SearchResult"},"type":"array","title":"Results"}},"type":"object","required":["results"],"title":"SearchResponse"},"SearchResult":{"properties":{"chapter_id":{"type":"string","title":"Chapter Id"},"chapter_title":{"type":"string","title":"Chapter Title"},"content":{"type":"string","title":"Content"},"similarity_score":{"type":"number","title":"Similarity Score"}},"type":"object","required":["chapter_id","chapter_title","content","similarity_score"],"title":"SearchResult"},"Source":{"properties":{"chapter_id":{"type":"string","title":"Chapter Id"},"chapter_title":{"type":"string","title":"Chapter Title"},"content":{"type":"string","title":"Content"},"similarity_score":{"type":"number","title":"Similarity Score"}},"type":"object","required":["chapter_id","chapter_title","content","similarity_score"],"title":"Source"},"TextbookChapter":{"properties":{"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"slug":{"type":"string","title":"Slug"},"order":{"type":"integer","title":"Order"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"},"content_blocks":{"anyOf":[{"items":{"$ref":"#/components/schemas/ContentBlock"},"type":"array"},{"type":"null"}],"title":"Content Blocks"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["title","content","slug","order","id","created_at","updated_at"],"title":"TextbookChapter"},"TextbookChapterWithId":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"slug":{"type":"string","title":"Slug"},"order":{"type":"integer","title":"Order"}},"type":"object","required":["id","title","slug","order"],"title":"TextbookChapterWithId"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}